rename form, fix story
