+++ components/EventCard/EventCard.stories.js
@@ -11,6 +11,7 @@ const Template = (args) => <EventCard {...args} />;
 export const DefaultEventCard = Template.bind({});
 DefaultEventCard.args = {
   event: {
+    id: "1234",
     name: "Training",
     date: "2022-12-02",
     startTime: "2022-12-02 20:00",
