+++ pages/events/[id]/index.js
@@ -18,6 +18,9 @@ function EventDetails() {
 
   const eventDate = formatDate(event.date);
   const eventTime = formatTime(event.startTime, event.endTime);
+  const eventLocation = !event.location ? "As usual" : event.location;
+  const eventDescription = !event.description ? "Let's goooo" : event.description;
+
   return (
     <>
       <Wrapper>
@@ -27,11 +30,11 @@ function EventDetails() {
           <time>{eventDate} </time>
           <time>{eventTime}</time>
         </p>
-        <p>Location: {event.location}</p>
-        <p>Description: {event.description}</p>
+        <p>Location: {eventLocation}</p>
+        <p>Description: {eventDescription}</p>
       </Wrapper>
 
-      <StyledLink href="" variant="fab">
+      <StyledLink href={`/events/${id}/edit`} variant="fab">
         <Svg variant="edit" />
       </StyledLink>
     </>
